requirements.txt
PyQt5==5.15.9
pandas==2.1.3
requests==2.31.0
smartapi-python==1.3.5
python-dotenv==1.0.0
cryptography==41.0.7
pytest==7.4.3
pytest-qt==4.2.0
logging==0.4.9.6 